javascript: prepare compatibility of Greasemonkey script with Chromium/Chrome
Chromium does not support GM_xmlHttpRequest. We circumvent this
limitation with the use of Access-Control-Allow-Origin policy in the
headers of our answers.

Also Chromium installs the redirection page instead of the user
script. The redirection page is renamed to /greasemonkey to ensure
that there is no confusion and redirects to ipoo.user.js resource.