collector: clean-up handling of cached deferred
When we cache a deferred object, its result depends on other callbacks
that have been attached to it, before we attach our own. Therefore,
when we cache the deferred object, the entity that attach its callback
first will be allowed to alter the result of the callback if we just
return the same deferred object to a later entity.

Therefore, we save the original deferred in some structure and the
cache system will return a new deferred for each entity requiring the
deferred object. Those new deferreds will be trigerred when the
original deferred is trigerred.